pyproject.toml
[project]
name = "absolute-zero"
description = "Your environment description here"
tags = ["placeholder-tag", "train", "eval"]
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "verifiers>=0.1.2.post1",
    "nest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["absolute_zero.py"]
